Updated Payment, Company and Post

+++ src/main/java/com/example/pathfinder/Data/PaymentData/Payment.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long paymentId;
+    private int paymentId;
     private double amount;
     private String paymentMethod;
     private String cardNumber;
